Update 3 - Connect4_project.c
I changed a few things. In the beginning, I just declared the functions and I wrote their definitions after the "main" function. I also made the "play against computer" option function, now we can play a game against a simple AI. But the board is still problematic, so we need to resolve that. Also, the AI is pretty stupid for now and it cannot do advanced moves, so you can win pretty easily.