Update 4 - Connect4_project.c
So firstly, today I upgraded the "getComputerMove" function, this is the AI which plays against the user.  In the last update the AI was really basic, just good enough for testing, but now It can make some better moves. As I discussed with my fried Felix, I just added many "if" statements, so the AI can make the best move every time. Also in the last update the board was really messy and buggy, but now It has a better design, and it is also not buggy anymore. But sometimes it is really hard to see the objects on the board, so we might change the board again in the future, or we will add colors to differentiate the player and the computer.